render page and test pass
