+++ src/RenderPokemon.js
@@ -1,4 +1,5 @@
 import React, { Component } from 'react'
+import { Link } from 'react-router-dom';
 import './RenderPokemon.css';
 
 export default class RenderPokemon extends Component {
@@ -6,7 +7,9 @@ export default class RenderPokemon extends Component {
         return (
             <li>
                 <div className="immageContainer">
+                    <Link to={`/PokemonDetail/${this.props.pokemon._id}`}>
                     <h2>{this.props.pokemon.pokemon}</h2>
+                    </Link>
                     <p>Pokemon type: {this.props.pokemon.type_1}</p>
                     <p>Attack levle: {this.props.pokemon.attack}</p>
                     <p>Defense levle: {this.props.pokemon.defense}</p>
