+++ src/RenderPokemon.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import App from './App';
+import App from './SearchPage';
 import Adapter from 'enzyme-adapter-react-16'
